コミット本社9/11

+++ S_management/src/main/java/com/example/demo/User4PK.java
@@ -4,11 +4,7 @@ import java.io.Serializable;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.Table;
 
 import lombok.Data;
 
@@ -19,7 +15,7 @@ public class User4PK implements Serializable {
 
 	@Id
     @Column(name = "id")
-    private long id;
+    private int id;
 
 	@Id
     @Column(name = "branchnum")
